loader: efi updates from freebsd

+++ usr/src/boot/sys/boot/efi/boot1/zfs_module.c
@@ -55,9 +55,9 @@ vdev_read(vdev_t *vdev, void *priv, off_t off, void *buf, size_t bytes)
 	status = devinfo->dev->ReadBlocks(devinfo->dev,
 	    devinfo->dev->Media->MediaId, lba, bytes, buf);
 	if (status != EFI_SUCCESS) {
-		DPRINTF("vdev_read: failed dev: %p, id: %u, lba: %zu, size: %zu,"
+		DPRINTF("vdev_read: failed dev: %p, id: %u, lba: %ju, size: %zu,"
                     " status: %lu\n", devinfo->dev,
-                    devinfo->dev->Media->MediaId, lba, bytes,
+                    devinfo->dev->Media->MediaId, (uintmax_t)lba, bytes,
                     EFI_ERROR_CODE(status));
 		return (-1);
 	}
